Including cleanup for mc17-0.9-bc7

+++ common/logisticspipes/network/PacketInboundHandler.java
@@ -1,7 +1,5 @@
 package logisticspipes.network;
 
-import io.netty.channel.ChannelHandlerContext;
-import io.netty.channel.SimpleChannelInboundHandler;
 import logisticspipes.LPConstants;
 import logisticspipes.LogisticsPipes;
 import logisticspipes.network.PacketHandler.InboundModernPacketWrapper;
@@ -9,22 +7,25 @@ import logisticspipes.network.exception.TargetNotFoundException;
 import logisticspipes.proxy.MainProxy;
 import logisticspipes.proxy.SimpleServiceLocator;
 
-import org.apache.logging.log4j.Level;
-
 import cpw.mods.fml.common.FMLLog;
 
+import io.netty.channel.ChannelHandlerContext;
+import io.netty.channel.SimpleChannelInboundHandler;
+import org.apache.logging.log4j.Level;
+
 public class PacketInboundHandler extends SimpleChannelInboundHandler<InboundModernPacketWrapper> {
+
 	@Override
 	protected void channelRead0(ChannelHandlerContext ctx, InboundModernPacketWrapper msg) throws Exception {
 		try {
 			msg.packet.processPacket(msg.player);
-			if(LPConstants.DEBUG) {
-				PacketHandler.debugMap.remove((Integer) msg.packet.getDebugId());
+			if (LPConstants.DEBUG) {
+				PacketHandler.debugMap.remove(msg.packet.getDebugId());
 			}
-		} catch(TargetNotFoundException e) {
-			if(msg.packet.retry() && MainProxy.isClient(msg.player.getEntityWorld())) {
+		} catch (TargetNotFoundException e) {
+			if (msg.packet.retry() && MainProxy.isClient(msg.player.getEntityWorld())) {
 				SimpleServiceLocator.clientBufferHandler.queueFailedPacket(msg.packet, msg.player);
-			} else if(LPConstants.DEBUG) {
+			} else if (LPConstants.DEBUG) {
 				LogisticsPipes.log.error(msg.packet.getClass().getName());
 				LogisticsPipes.log.error(msg.packet.toString());
 				e.printStackTrace();
